Address code review comments
- Move ExceptionBlock to it own class
- Remove unused code for storedStack
- Create a method to calculate list index from stack index
- Make JumpUtils constructor private
